requirements.txt
sentence-transformers
faiss-cpu
numpy
tqdm
python-dotenv
groq
fastapi
uvicorn
